Multiple things: file split, comments, major bug..
Split the moisture_sensors .h and .cpp files into one .h and .cpp file
for each type of moisture sensor. Having them all lumped into one file
would get messy very quickly... I may do a similar thing to the other
libraries.

Added many lines of comments to the SIP.ino file and restructured the
preprocessor commands to make things more clear.

In testing the file split, I found and fixed a major bug where the test
sensor (a potentiometer) was giving bogus values that drifted for no
obvious reason. The issue is that the pin was never set in the
constructor of the TestSensor class.

In fixing the bug, I added code in the SensorArduino class to print out
the sensor values to the computer.