Added functionality for the STTDataPoint.
Instead of making a new STTDataPoint every time a measurement is made,
the moisture sensor edits the STTDataPoint's already in the DataStream.
The cyclic ArrayDataStream works nearly the same, just uses the
manually_update_next function instead of the add_data function. When the
moisture sensor creates the data stream, the moisture sensor fills the
data stream with empty STTDataPoint objects.